womens_college_basketball_repo/requirements.txt
# Women's College Basketball Betting System Requirements

# Core dependencies
python>=3.8

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# API and networking
requests>=2.26.0
aiohttp>=3.8.0
asyncio

# Configuration
pydantic>=1.9.0
python-dotenv>=0.19.0

# AI/ML (optional but recommended)
anthropic>=0.3.0
openai>=0.27.0

# Database (optional)
sqlalchemy>=1.4.0

# Logging and monitoring
loguru>=0.6.0

# Testing (development)
pytest>=7.0.0
pytest-asyncio>=0.18.0

# Utilities
pytz>=2021.3
dateutil>=2.8.2
